LinearMixingModels

Documentation for LinearMixingModels.

LinearMixingModels.ILMM

LinearMixingModels.IndependentMOGP

LinearMixingModels.OILMM

LinearMixingModels.isotopic_inputs

Base.rand

LinearMixingModels.denoised_marginals

LinearMixingModels.finite_gps

LinearMixingModels.finite_gps

LinearMixingModels.independent_mogp

LinearMixingModels.rand_latent

LinearMixingModels.isotopic_inputs — Type

A constant to represent all isotopic input types.

LinearMixingModels.ILMM — Type

ILMM(fs, H, D)

An Instantaneous Linear Mixing Model (ILMM) – a distribution over vector- valued functions. Let p be the number of observed outputs, and m the number of latent processes, then

Arguments:

fs: a length-m vector of Gaussian process objects from AbstractGPs.jl.

H: a p x m matrix representing a fixed basis of our p-dim target: h1,...,hm

D: an m x m Diagonal matrix, variance of noise on each latent process. Positive entries.

LinearMixingModels.IndependentMOGP — Type

IndependentMOGP

LinearMixingModels.OILMM — Type

OILMM(fs, H, D)

An Orthogonal Instantaneous Linear Mixing Model (OILMM) – a distribution over vector- valued functions. Let p be the number of observed outputs, and m the number of latent processes, then

Arguments:

fs: a length-m vector of Gaussian process objects as an IndependentMOGP.

H: a p x m orthogonal matrix representing a fixed basis of our p-dim target: h1,...,hm

D: an m x m Diagonal matrix, variance of noise on each latent process. Positive entries.

Base.rand — Method

rand(rng::AbstractRNG, fx::FiniteGP{<:OILMM})

Sample from the OILMM, including the observation noise. Follows generative structure of model 2 from [1]. Follows the AbstractGPs.jl API. See also rand_latent. [1] - Bruinsma et al 2020.

LinearMixingModels.denoised_marginals — Method

denoised_marginals(fx::FiniteGP{<:OILMM})

Returns the marginal distribution over the OILMM without the IID noise components. See also marginals.

LinearMixingModels.finite_gps — Method

finite_gps()

Returns a list of of the finite GPs for all latent processes for some input x.

LinearMixingModels.finite_gps — Method

finite_gps()

Returns a list of of the finite GPs for all latent processes.

LinearMixingModels.independent_mogp — Method

independent_mogp

Returns an IndependentMOGP given a list of AbstractGPs.

LinearMixingModels.rand_latent — Method

rand_latent(rng::AbstractRNG, fx::FiniteGP{<:OILMM})

Sample from the latent (noiseless) process. See also rand.